+++ backend/payments/apps.py
@@ -5,7 +5,7 @@ class PaymentsConfig(AppConfig):
     name = "payments"
 
     def ready(self):
-        import payments.converter  # noqa
+        import api.payments.converter  # noqa
 
         from payments.providers.stripe import Stripe
