+++ backend/tests/voting/test_types.py
@@ -1,4 +1,4 @@
-from voting.types import VoteValues
+from api.voting.types import VoteValues
 
 
 def test_get_enum_member_from_int():
