+++ backend/voting/apps.py
@@ -5,4 +5,4 @@ class VotingConfig(AppConfig):
     name = "voting"
 
     def ready(self):
-        import voting.converter  # noqa
+        import api.voting.converter  # noqa
